requirements.txt
# Core dependencies
ccxt>=4.2.0
web3>=6.15.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Terminal UI
rich>=13.7.0

# Utilities
python-dotenv>=1.0.0
